test:Frames 인스턴스 생성 여부 테스트작성
